+++ lib/serif/site.rb
@@ -268,14 +268,32 @@ class Site
       end
     end
 
-    posts.each do |post|
+    # the posts are iterated over in reverse chrological order
+    next_post = nil
+
+    # run through the posts + nil so we can keep |a, b| such that a hits every element
+    # while iterating.
+    [*posts, nil].each_cons(2) do |post, prev_post|
       puts "Processing post: #{post.path}"
 
       FileUtils.mkdir_p(tmp_path(File.dirname(post.url)))
 
       File.open(tmp_path(post.url + ".html"), "w") do |f|
-        f.puts default_layout.render!("site" => self, "page" => { "title" => ["Posts", "#{post.title}"] }, "content" => Liquid::Template.parse(File.read("_templates/post.html")).render!("post" => post))
+        # variables available in the post template
+        post_template_variables = {
+          "post" => post,
+          "prev_post" => prev_post,
+          "next_post" => next_post
+        }
+
+        f.puts default_layout.render!(
+          "site" => self,
+          "page" => { "title" => ["Posts", "#{post.title}"] },
+          "content" => Liquid::Template.parse(File.read("_templates/post.html")).render!(post_template_variables)
+        )
       end
+
+      next_post = post
     end
 
     generate_archives(default_layout)