+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # next release
 
-* ...
+* Make the previous and next posts available from individual post pages. (#3)
 
 # v0.2.1